Extract duplicate section scanner
