+++ src/scan/mod.rs
@@ -1,5 +1,6 @@
 //! The probes module contains the code that scans for problems in a `TikiBase`
 //! and reports problems found as `Issues`.
 
+pub(crate) mod duplicate_sections;
 pub(crate) mod section_capitalization;
 pub(crate) mod section_level;
